Fix Issue #16 - Change openpomodoro client init
The --directory flag does not work, all commands use the default
`~/.pomodoro` directory even when the flag is provided.
With the current code, `directoryFlag` is always an empty string when
passed to the openpomodoro.NewClient function because the flag has not
been parsed yet at this time.
Moving to the PersistentPreRun hook ensures that all commands will
initialize the client with the correct flag value.

Co-authored-by: hugobally <hugobally@users.noreply.github.com>